Move staking tests to separate file
